+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features
 
 -   Add support for monotonic constraints.
+-   Add support for Mean average error (MAE) loss.
+-   Add support for Poisson loss.
 
 ## 1.6.0 2023-09-27